+++ scripts/run_checks.py
@@ -42,7 +42,10 @@ COMPILE_TARGETS = (
     "goldmonitor/desktop_notification_runtime.py",
     "goldmonitor/notification_adapters.py",
     "goldmonitor/notification_channel_runtime.py",
+    "goldmonitor/notification_delivery.py",
+    "goldmonitor/notification_policy.py",
     "goldmonitor/notification_runtime.py",
+    "goldmonitor/notification_transport.py",
     "goldmonitor/operations_runtime.py",
     "goldmonitor/data_archive.py",
     "goldmonitor/data_archive_runtime.py",
